Start libvirt domains once created
Previously libvirt domains where only started after creation if
autostart had been enabled as well. In case autostart was not
enabled, a handler waited for the libvirt domain to setup,
reboot and become usable although it was never started.

Now, libvirt domains are started if they should be present and
its domain state is set to running.