Added missing else section to inline if-expression in cloudinit_networkconfig variable
Task 'Fail if required vars are not set' in role jm1.cloudy.libvirt_server
evaluates variable cloudinit_userdata which uses cloudinit_networkconfig.
Variable cloudinit_networkconfig evaluates libvirt_netplan which could be
an empty dictionary in which case the inline if-expression evaluates to
false and then fails to evaluate the missing else section.