Drop Netplan install from role jm1.cloudy.netplan
Netplan and role jm1.cloudy.netplan might not be used and role variable
netplan_config might not be set, so do not install Netplan preemptively.
On systems which do not have Netplan installed yet it is likely that
additional work has to be done such as deactivating the previous
network configuration mechanism.