Moved role jm1.cloudy.netplan below role jm1.cloudy.packages etc.
On systems without Netplan it has to be installed first and probably
additional work has to be done such as deactivating the previous
network configuration mechanism. So role jm1.cloudy.netplan has
to be run later so that e.g. Netplan can be installed before with
role jm1.cloudy.packages and services can be deactivated with
jm1.cloudy.services etc.