Refactored roles related to OpenStack Nova (Compute) instances
Replaced most openstack_server_* role variables for server creation
with single openstack_server_config variable following conventions
of most other roles in this collection.

Renamed openstack_network_ports into openstack_server_ports to
highlight its belonging to the openstack_server role and to match
variable naming conventions.

Variable openstack_server_floating_ips is still listed as a unique
variable because its format is different from parameter floating_ips
of module openstack.cloud.server.

Variable openstack_server_name is still listed as a unique variable
because it is used in other roles as well and is most likely to be
changed which is easier if openstack_server_config has not to be
cloned for the sake of renaming a compute instance. Its default
value has been changed from fqdn to inventory_hostname because
fqdn is not defined by Ansible.

Variable openstack_server_state has been added to role defaults
because it is used by the role.