Fixed conditional which broke when previous task where skipped
All tasks in a block inherit directives applied at the block level. The
directive does not affect the block itself, it is only inherited by the
tasks enclosed by a block [1]. If a task fails or is skipped, Ansible
still registers a variable with a failure or skipped status, unless the
task is skipped based on tags [2].

Suppose task "Reload domain info" is skipped because the conditional
inherited from the surrounding block evaluated to false, for example
variable libvirt_domain_state is set to shutoff and
libvirt_domain_info[libvirt_domain].state evaluates to shutdown. Then
variable libvirt_domain_info will be registered with a skipped status.
Next, the conditional of task "Destroy domain" is evaluated. Its
evaluation is not skipped because the when directive which is applied at
the block level does not affect the block itself but is inherited by the
tasks enclosed by the block. Ansible will fail during evaluation of this
conditional inherited from the block:

  libvirt_domain_info[libvirt_domain].state != 'shutdown'

Variable libvirt_domain_info has been registered with a skipped status
in the previous task and thus libvirt_domain_info does not have an
attribute identified by libvirt_domain, so the conditional breaks with:

 'dict object' has no attribute '...'

To fix this, we test whether the previous task has been skipped. This
extra conditional has to be added to the surrounding block because we
want to prevent Ansible from evaluating

  libvirt_domain_info[libvirt_domain].state != 'shutdown'

when libvirt_domain_info has been registered with a skipped status.
The conditional from task "Destroy domain" can be removed safely
because this conditional is inherited from the surrouding block.

Ref.:
[1] https://docs.ansible.com/ansible/latest/user_guide/playbooks_blocks.html
[2] https://docs.ansible.com/ansible/latest/user_guide/playbooks_variables.html