Added task to enable and start hwfp's systemd service
Previously, hwfp's systemd service was enabled and started in a handler,
but when playbook execution failed, then it was possible that this
handler would have never been executed. To ensure that the service is
always enabled and started an extra task has been added. The handler
has been reduced to restart the service on changes only.