Added role jm1.cloudy.openshift_ipi for creating OpenShift and OKD clusters with IPI
Split services section in plabooks/site.yml into early and late
services. For example, Ansible host *-430-okd-ipi-provisioner
has to run jm1.libvirt.setup first to install libvirtd and then
run jm1.cloudy.openshift_ipi later because the latter connects
to libvirtd to launch its bootstrap virtual machine.

Added Ansible group bootstrap_only which will cause hosts to be
created and booted but not configured with Ansible.

Updated entrypoint.sh to launch ssh-agent in containerized setup
and added instructions to README.md on how to ensure it is running
on bare-metal systems. This is a prerequisite for SSH agent
forwarding which is required to access the nested bootstrap virtual
machine which is created by OpenShift Installer-provisioned
installation (IPI).