Fixed when keyword and dropped register keyword from *_config variables
Each *_config variable is a list of tasks whose items a Ansible loop
will feed to module jm1.ansible.execute_module. Because of this loop,
the register keyword in *_config variables will not work the same way
as Ansible's register keyword being applied to a single (non-loop)
task. Keyword register in *_config variables is in fact useless.