Skip collection install for jm1.cloudy in container entrypoint if it is available
For example, this skips the installation from Ansible Galaxy when the
collection is mounted into the container already. This allows
developers to test changes to the collection without the ansible-galaxy
command to override the mounted collection by downloading it to again
from Ansible Galaxy to a different location with higher precedence.