Run Ansible playbook setup.yml as cloudy user
When running as root user, Ansible will search for collections relative
to root's paths instead of cloudy's paths. For example, when
ansible.cfg.example is used then collections will be installed in
cloudy's home, but when running Ansible as root it will search in root's
home causing Ansible to fail.