Handle interrupts and exit codes properly in container entrypoint
Previously, the trap would kill the whole shell process group at the
end of the entrypoint execution. This would cause the script to have an
exit code which did not reflect the last status of the last entrypoint
command but of the kill command of the trap. Now, the script will no
longer install the trap for EXIT signals, which was unnecessary anyway,
and return the exit code of the last entrypoint command.