Undefine openshift_tests_pullsecret to fix None conditionals
The previous default value for openshift_tests_pullsecret was
interpreted by Ansible 2.10.8 as an empty string. This broke the None
conditionals in jm1.cloudy.openshift_tests. The latter would write an
empty pull-secret.txt file which would be passed to
'oc adm release info' and causing it to fail with:

  error: unable to load --registry-config: reading JSON file
  "/home/ansible/pull-secret.txt": unmarshaling JSON at
  "/home/ansible/pull-secret.txt": json: cannot unmarshal string into
  Go value of type config.dockerConfigFile

This change undefines openshift_tests_pullsecret as it is not required
for OKD releases anyway.